backend/requirements.txt
django
djangorestframework
pandas
numpy
prophet
openpyxl
